redis::cluster

Struct ClusterClient

pub struct ClusterClient { /* private fields */ }

Available on crate feature cluster only.

A Redis Cluster client, used to create connections.

Implementations

impl ClusterClient

pub fn new<T: IntoConnectionInfo>( initial_nodes: impl IntoIterator<Item = T>, ) -> RedisResult<ClusterClient>

Creates a ClusterClient with the default parameters.

This does not create connections to the Redis Cluster, but only performs some basic checks on the initial nodesโ URLs and passwords/usernames.

Errors

Upon failure to parse initial nodes or if the initial nodes have different passwords or usernames, an error is returned.

pub fn builder<T: IntoConnectionInfo>( initial_nodes: impl IntoIterator<Item = T>, ) -> ClusterClientBuilder

Creates a ClusterClientBuilder with the provided initial_nodes.

pub fn get_connection(&self) -> RedisResult<ClusterConnection>

Creates new connections to Redis Cluster nodes and returns a cluster::ClusterConnection.

Errors

An error is returned if there is a failure while creating connections or slots.

pub fn get_connection_with_config( &self, config: ClusterConfig, ) -> RedisResult<ClusterConnection>

Creates new connections to Redis Cluster nodes with a custom config and returns a cluster_async::ClusterConnection.

Errors

An error is returned if there is a failure while creating connections or slots.

pub async fn get_async_connection(&self) -> RedisResult<ClusterConnection>

Available on crate feature cluster-async only.

Creates new connections to Redis Cluster nodes and returns a cluster_async::ClusterConnection.

Errors

An error is returned if there is a failure while creating connections or slots.

pub async fn get_async_connection_with_config( &self, config: ClusterConfig, ) -> RedisResult<ClusterConnection>

Available on crate feature cluster-async only.

Creates new connections to Redis Cluster nodes with a custom config and returns a cluster_async::ClusterConnection.

Errors

An error is returned if there is a failure while creating connections or slots.

pub fn get_pending_async_connection_with_config( &self, config: ClusterConfig, ) -> ClusterConnection

Available on crate feature cluster-async only.

Creates new connections to Redis Cluster nodes with a custom config and returns a cluster_async::ClusterConnection. The connections to the cluster nodes are done in the background so the caller wonโt know if the cluster is available until the first command is sent.