redis

Module cluster_routing

Available on crate feature cluster only.

Routing information for cluster commands. Routing information for cluster commands.

Structs

Route

Defines the slot and the SlotAddr to which a command should be sent

Enums

AggregateOp

Numerical aggregating operators.

LogicalAggregateOp

Logical bitwise aggregating operators.

MultiSlotArgPattern

Represents the pattern of argument structures in multi-slot commands, defining how the arguments are organized in the command.

MultipleNodeRoutingInfo

Defines which collection of nodes should receive a request

ResponsePolicy

Policy defining how to combine multiple responses into one.

RoutingInfo

Defines whether a request should be routed to a single node, or multiple ones.

SingleNodeRoutingInfo

Defines which single node should receive a request.

SlotAddr

What type of node should a request be routed to, assuming read from replica is enabled.