redis

Module vector_sets

Available on crate feature vector-sets only.

Defines types to use with the vector sets commands.

Structs

VAddOptions

Options for the VADD command

VEmbOptions

Options for the VEMB command

VSimOptions

Options for the VSIM command

Enums

EmbeddingInput

Input data formats that can be used to generate vector embeddings:

VectorAddInput

Represents different ways to input data for vector add commands

VectorQuantization

Quantization options for vector storage

VectorSimilaritySearchInput

Represents different ways to input query data for vector similarity search commands